+++ docs/versionhistory.rst
@@ -3,6 +3,10 @@ Version history
 
 This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
+**UNRELEASED**
+
+- Added asynchronous ``Path`` class
+
 **3.2.1**
 
 - Fixed idle thread pruning on asyncio sometimes causing an expired worker thread to be assigned a